python_trader/requirements.txt
# MetaTrader 5 Python Bot Requirements
MetaTrader5>=5.0.45
pandas>=1.5.0
numpy>=1.21.0
flask>=2.3.0
flask-cors>=3.0.10
requests>=2.28.0
python-dotenv>=1.0.0
